backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
openai==1.3.0
twilio==8.10.0
pymongo==4.6.0
python-dotenv==1.0.0
